+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Update broken documentation links in readme.
 * Set minimum php-sdk version [1.1.4.2](https://github.com/unzerdev/php-sdk/releases/tag/1.1.4.2).
 * Change translation keys of invoice payment methods to avoid translation conflicts with shop system.
+* Display Module version in Backend configuration.
 * Several minor improvements.
 
 ### Fix